requirements.txt
dash
dash-auth
dash-bootstrap-components<1
dash-core-components
dash-html-components
dash-table
Flask
Jinja2
numpy
openpyxl
pandas
plotly
mysql-connector-python
scikit-learn
matplotlib
tensorflow
keras
python-dotenv
seaborn
xgboost